oas20: extract request handler to a helper method
